Rename to backup-tracker (CORE-848)
Renames from backup-manager to backup-tracker to better reflect actual
intent of the module and API
