+++ backup-tracker/src/main/java/io/telicent/smart/cache/backups/BackupTransitionListener.java
@@ -1,0 +1,32 @@
+/**
+ * Copyright (C) Telicent Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.telicent.smart.cache.backups;
+
+/**
+ * Listener for backup state transitions
+ */
+@FunctionalInterface
+public interface BackupTransitionListener {
+
+    /**
+     * Called when a {@link SimpleBackupTracker} derived tracker transitions between states
+     *
+     * @param tracker  Tracker instance
+     * @param oldState Old state
+     * @param newState New state
+     */
+    void accept(BackupTracker tracker, BackupTrackerState oldState, BackupTrackerState newState);
+}
