+++ backup-tracker/src/test/java/io/telicent/smart/cache/backups/kafka/DockerTestKafkaBackupTracker.java
@@ -1,0 +1,64 @@
+/**
+ * Copyright (C) Telicent Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.telicent.smart.cache.backups.kafka;
+
+import io.telicent.smart.cache.backups.BackupTracker;
+import io.telicent.smart.cache.backups.BackupTrackerState;
+import io.telicent.smart.cache.sources.kafka.KafkaEventSource;
+import io.telicent.smart.cache.sources.kafka.sinks.KafkaSink;
+import org.testng.annotations.*;
+
+import java.util.List;
+import java.util.UUID;
+import java.util.function.Consumer;
+
+public class DockerTestKafkaBackupTracker extends AbstractKafkaBackupTrackerTests {
+
+
+    @Test(dataProvider = "transitions")
+    public void givenKafkaBackupTrackers_whenTransitioningStates_thenStateTrackersAreSynced(
+            List<Consumer<BackupTracker>> transitions, BackupTrackerState expectedFinalState) {
+        // Given
+        try (KafkaSink<UUID, BackupTransition> sink = createSink()) {
+            KafkaEventSource<UUID, BackupTransition> source = createSource(
+                    "test-backup-tracker" + GROUP_ID.incrementAndGet());
+
+            try (BackupTracker primary = KafkaPrimaryBackupTracker.builder().application("test").sink(sink).build()) {
+                verifySecondaryTracker("test", source, transitions, primary, expectedFinalState);
+            }
+        }
+    }
+
+    @Test(dataProvider = "splitTransitions")
+    public void givenKafkaBackupTrackers_whenTransitioningStatesAndSecondaryRestarts_thenStateTrackersAreSynced(
+            List<Consumer<BackupTracker>> preTransitions, List<Consumer<BackupTracker>> postTransitions,
+            BackupTrackerState expectedInterimState, BackupTrackerState expectedFinalState) {
+        // Given
+        String groupId = "test-backup-tracker" + GROUP_ID.incrementAndGet();
+        try (KafkaSink<UUID, BackupTransition> sink = createSink()) {
+            try (BackupTracker primary = KafkaPrimaryBackupTracker.builder().application("test").sink(sink).build()) {
+                // When
+                verifySecondaryTracker("test", createSource(groupId), preTransitions, primary, expectedInterimState);
+
+                // Then
+                // We close and restart the secondary applying further transitions
+                // NB - We reuse the same groupId so should resume consumption from previously committed offsets, if any
+                verifySecondaryTracker("test", createSource(groupId), postTransitions, primary, expectedFinalState);
+            }
+        }
+    }
+
+}
