+++ cli/cli-core/src/test/java/io/telicent/smart/cache/cli/commands/backup/BackupSecondary.java
@@ -1,12 +1,24 @@
+/**
+ * Copyright (C) Telicent Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package io.telicent.smart.cache.cli.commands.backup;
 
-import com.github.rvesse.airline.annotations.AirlineModule;
 import com.github.rvesse.airline.annotations.Command;
-import io.telicent.smart.cache.backups.BackupManager;
-import io.telicent.smart.cache.backups.BackupManagerState;
-import io.telicent.smart.cache.cli.commands.SmartCacheCommand;
-import io.telicent.smart.cache.cli.options.BackupManagerOptions;
-import io.telicent.smart.cache.cli.options.KafkaConfigurationOptions;
+import io.telicent.smart.cache.backups.BackupTracker;
+import io.telicent.smart.cache.backups.BackupTrackerState;
+import io.telicent.smart.cache.backups.BackupTransitionListener;
 import lombok.AllArgsConstructor;
 import lombok.NonNull;
 import lombok.ToString;
@@ -14,7 +26,6 @@ import lombok.ToString;
 import java.time.Duration;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.function.BiConsumer;
 
 @Command(name = "secondary", description = "Runs the secondary backup service")
 public class BackupSecondary extends BackupPrimary {
@@ -23,7 +34,7 @@ public class BackupSecondary extends BackupPrimary {
     public int run() {
         AtomicBoolean paused = new AtomicBoolean(false);
         long start = System.currentTimeMillis();
-        try (BackupManager secondary = this.backupManagerOptions.getSecondary(null, BackupPrimary.APP_ID,
+        try (BackupTracker secondary = this.backupTrackerOptions.getSecondary(null, BackupPrimary.APP_ID,
                                                                               this.kafkaOptions,
                                                                               BackupPrimary.APP_ID,
                                                                               List.of(new PauseListener(paused)))) {
@@ -53,13 +64,13 @@ public class BackupSecondary extends BackupPrimary {
 
     @AllArgsConstructor
     @ToString
-    private static final class PauseListener implements BiConsumer<BackupManagerState, BackupManagerState> {
+    private static final class PauseListener implements BackupTransitionListener {
 
         @NonNull
         private final AtomicBoolean paused;
 
         @Override
-        public void accept(BackupManagerState from, BackupManagerState to) {
+        public void accept(BackupTracker tracker, BackupTrackerState from, BackupTrackerState to) {
             switch (to) {
                 case BACKING_UP, RESTORING:
                     this.paused.set(true);
