+++ cli/cli-debug/src/test/java/io/telicent/smart/cache/cli/commands/debug/AbstractDockerDebugCliTests.java
@@ -32,7 +32,6 @@ import io.telicent.smart.cache.projectors.driver.ProjectorDriver;
 import io.telicent.smart.cache.projectors.sinks.events.file.EventCapturingSink;
 import io.telicent.smart.cache.sources.Event;
 import io.telicent.smart.cache.sources.EventHeader;
-import io.telicent.smart.cache.sources.Header;
 import io.telicent.smart.cache.sources.file.FileEventFormatProvider;
 import io.telicent.smart.cache.sources.file.FileEventFormats;
 import io.telicent.smart.cache.sources.file.FileEventSource;
