+++ projectors-core/src/main/java/io/telicent/smart/cache/projectors/sinks/AbstractTransformingSink.java
@@ -17,7 +17,6 @@ package io.telicent.smart.cache.projectors.sinks;
 
 import io.telicent.smart.cache.projectors.Sink;
 import io.telicent.smart.cache.projectors.SinkException;
-import lombok.ToString;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
