+++ projectors-core/src/test/java/io/telicent/smart/cache/projectors/sinks/TestFilterSink.java
@@ -25,7 +25,6 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 
 import java.io.IOException;
-import java.nio.file.FileSystem;
 import java.time.Duration;
 import java.util.*;
 import java.util.function.Predicate;
